requirements.txt
yfinance
pandas
google-generativeai
python-dotenv
pydantic
openai
